Move ArrayIndexer class to core.
