+++ src/axom/core/tests/CMakeLists.txt
@@ -17,6 +17,7 @@
 set(core_serial_tests
     core_about.hpp
     core_array.hpp
+    core_array_indexer.hpp
     core_array_for_all.hpp
     core_utilities.hpp
     core_bit_utilities.hpp
