+++ src/axom/quest/CMakeLists.txt
@@ -59,7 +59,6 @@ set( quest_headers
 
     util/mesh_helpers.hpp
     MeshViewUtil.hpp
-    ArrayIndexer.hpp
     )
 
 set( quest_sources
