+++ src/axom/quest/examples/quest_marching_cubes_example.cpp
@@ -27,7 +27,7 @@
 #include "axom/primal.hpp"
 #include "axom/mint/mesh/UnstructuredMesh.hpp"
 #include "axom/mint/execution/internal/structured_exec.hpp"
-#include "axom/quest/ArrayIndexer.hpp"
+#include "axom/core/ArrayIndexer.hpp"
 #include "axom/quest/MarchingCubes.hpp"
 #include "axom/quest/MeshViewUtil.hpp"
 #include "axom/sidre.hpp"
